requirements.txt
gpiozero
rpi-lgpio
luma.core
luma.led_matrix

